@@ -1,12 +1,31 @@
 module github.com/cybozu-go/well
 
+go 1.19
+
 require (
-	github.com/BurntSushi/toml v1.2.0
-	github.com/cybozu-go/log v1.6.1
-	github.com/cybozu-go/netutil v1.4.3
+	github.com/BurntSushi/toml v1.2.1
+	github.com/cybozu-go/log v1.7.0
+	github.com/cybozu-go/netutil v1.4.4
 	github.com/spf13/pflag v1.0.5
-	github.com/spf13/viper v1.12.0
-	golang.org/x/net v0.0.0-20220826154423-83b083e8dc8b
+	github.com/spf13/viper v1.15.0
+	golang.org/x/net v0.5.0
 )
 
-go 1.13
+require (
+	github.com/fsnotify/fsnotify v1.6.0 // indirect
+	github.com/hashicorp/hcl v1.0.0 // indirect
+	github.com/magiconair/properties v1.8.7 // indirect
+	github.com/mitchellh/mapstructure v1.5.0 // indirect
+	github.com/onsi/gomega v1.25.0 // indirect
+	github.com/pelletier/go-toml/v2 v2.0.6 // indirect
+	github.com/spf13/afero v1.9.3 // indirect
+	github.com/spf13/cast v1.5.0 // indirect
+	github.com/spf13/jwalterweatherman v1.1.0 // indirect
+	github.com/subosito/gotenv v1.4.2 // indirect
+	github.com/vishvananda/netlink v1.2.1-beta.2 // indirect
+	github.com/vishvananda/netns v0.0.3 // indirect
+	golang.org/x/sys v0.4.0 // indirect
+	golang.org/x/text v0.6.0 // indirect
+	gopkg.in/ini.v1 v1.67.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
